refactor(gh-design): extract filterReviewers to side-effect-free module, fix guard condition
- Move filterReviewers() into its own module to avoid DB side effects
  in tests (Copilot review comment)
- Fix guard condition: use filtered newReviewers.length instead of
  comparing against unfiltered REQUEST_REVIEWERS, which caused the
  task to re-enter and re-save on every 5-minute tick when the author
  was in the reviewer list (Copilot review comment)

Amp-Thread-ID: https://ampcode.com/threads/T-019ddde1-db08-71ca-809d-fd7e51b68842
